+++ src/scripting/ui/lua_tab_panel.h
@@ -16,33 +16,45 @@
  *
  */
 
-#ifndef WL_SCRIPTING_LUA_UI_H
-#define WL_SCRIPTING_LUA_UI_H
+#ifndef WL_SCRIPTING_UI_LUA_TAB_PANEL_H
+#define WL_SCRIPTING_UI_LUA_TAB_PANEL_H
 
-#include "scripting/lua.h"
-#include "scripting/luna.h"
-
-namespace UI {
-class Panel;
-}
+#include "scripting/ui/lua_panel.h"
+#include "ui_basic/tabpanel.h"
 
 namespace LuaUi {
 
-/*
- * Base class for all classes in wl.ui
- */
-class LuaUiModuleClass : public LunaClass {
+class LuaTabPanel : public LuaPanel {
 public:
-	const char* get_modulename() override {
-		return "ui";
+	LUNA_CLASS_HEAD(LuaTabPanel);
+
+	LuaTabPanel() = default;
+	explicit LuaTabPanel(UI::Panel* p) : LuaPanel(p) {
+	}
+	explicit LuaTabPanel(lua_State* L) : LuaPanel(L) {
+	}
+	~LuaTabPanel() override = default;
+
+	/*
+	 * Properties
+	 */
+	int get_no_of_tabs(lua_State* L);
+	int get_active(lua_State* L);
+	int set_active(lua_State* L);
+
+	/*
+	 * Lua Methods
+	 */
+	int remove_last_tab(lua_State* L);
+
+	/*
+	 * C Methods
+	 */
+	UI::TabPanel* get() {
+		return dynamic_cast<UI::TabPanel*>(panel_);
 	}
 };
 
-int upcasted_panel_to_lua(lua_State* L, UI::Panel* panel);
-
-std::string shortcut_string_if_set(const std::string& name, bool rt_escape);
-
-void luaopen_wlui(lua_State*, bool game_or_editor);
 }  // namespace LuaUi
 
-#endif  // end of include guard: WL_SCRIPTING_LUA_UI_H
+#endif  // end of include guard: WL_SCRIPTING_UI_LUA_TAB_PANEL_H
