+++ src/scripting/ui/lua_textarea.h
@@ -16,33 +16,43 @@
  *
  */
 
-#ifndef WL_SCRIPTING_LUA_UI_H
-#define WL_SCRIPTING_LUA_UI_H
+#ifndef WL_SCRIPTING_UI_LUA_TEXTAREA_H
+#define WL_SCRIPTING_UI_LUA_TEXTAREA_H
 
-#include "scripting/lua.h"
-#include "scripting/luna.h"
-
-namespace UI {
-class Panel;
-}
+#include "scripting/ui/lua_panel.h"
+#include "ui_basic/textarea.h"
 
 namespace LuaUi {
 
-/*
- * Base class for all classes in wl.ui
- */
-class LuaUiModuleClass : public LunaClass {
+class LuaTextarea : public LuaPanel {
 public:
-	const char* get_modulename() override {
-		return "ui";
+	LUNA_CLASS_HEAD(LuaTextarea);
+
+	LuaTextarea() = default;
+	explicit LuaTextarea(UI::Panel* p) : LuaPanel(p) {
+	}
+	explicit LuaTextarea(lua_State* L) : LuaPanel(L) {
+	}
+	~LuaTextarea() override = default;
+
+	/*
+	 * Properties
+	 */
+	int get_text(lua_State* L);
+	int set_text(lua_State* L);
+
+	/*
+	 * Lua Methods
+	 */
+
+	/*
+	 * C Methods
+	 */
+	UI::Textarea* get() {
+		return dynamic_cast<UI::Textarea*>(panel_);
 	}
 };
 
-int upcasted_panel_to_lua(lua_State* L, UI::Panel* panel);
-
-std::string shortcut_string_if_set(const std::string& name, bool rt_escape);
-
-void luaopen_wlui(lua_State*, bool game_or_editor);
 }  // namespace LuaUi
 
-#endif  // end of include guard: WL_SCRIPTING_LUA_UI_H
+#endif  // end of include guard: WL_SCRIPTING_UI_LUA_TEXTAREA_H
